requirements.txt
alembic~=1.13.1
annotated-types==0.6.0
anyio==4.2.0
cachetools==5.3.2
click==8.1.7
fastapi~=0.109.2
google-auth==2.28.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic~=2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
PyYAML==6.0.1
rsa==4.9
sniffio==1.3.0
SQLAlchemy~=2.0.27
starlette==0.36.3
typing_extensions==4.9.0
tzdata==2024.1
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
